Reorder scheduler packages
- Scheduler implemenation -> components
- types -> toplevel scheduler package
- integration_test.go -> integration
