+++ contrib/mesos/pkg/scheduler/components/podstoreadapter.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package scheduler
+package components
 
 import (
 	"k8s.io/kubernetes/contrib/mesos/pkg/queue"
