+++ contrib/mesos/pkg/scheduler/scheduler_mock.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package types
+package scheduler
 
 import (
 	"sync"
